refactor(metadata): centralize enrich IPC payload types in @shiranami/contracts
EnrichTrackInput / EnrichTrackResult / EnrichProgress / MetadataLookupResult
were duplicated across the main handler, preload bridge, and renderer store.
Move them into packages/contracts and wire the renderer to consume the
package (vite alias + tsconfig path + workspace dep), so the wire shape is
defined once.
