+++ apps/desktop/src/main/preload/api/metadata.ts
@@ -1,55 +1,21 @@
 import { ipcRenderer } from 'electron';
-import { IPC_CHANNELS } from '@shiranami/contracts';
+import {
+  IPC_CHANNELS,
+  type EnrichTrackInput,
+  type EnrichTrackResult,
+  type EnrichProgress,
+  type MetadataLookupResult,
+} from '@shiranami/contracts';
 import { createIpcListener } from '../ipc-listener';
 
 const C = IPC_CHANNELS.metadata;
 
-interface EnrichInputTrack {
-  id: string;
-  filePath: string;
-  title: string;
-  artist: string;
-  album: string;
-  albumArt: string | null;
-  genre: string;
-  year: number | null;
-  trackNumber: number | null;
-}
-
-interface EnrichResult {
-  id: string;
-  success: boolean;
-  updatedFields: Partial<{
-    title: string;
-    artist: string;
-    album: string;
-    genre: string;
-    year: number;
-    trackNumber: number;
-    albumArt: string;
-  }>;
-  source: string;
-  error?: string;
-}
-
-interface LookupResult {
-  title?: string;
-  artist?: string;
-  album?: string;
-  genre?: string;
-  year?: number;
-  trackNumber?: number;
-  coverImageUrl?: string;
-  source: 'itunes' | 'youtube' | 'none';
-  confidence: number;
-}
-
 export interface MetadataApi {
-  lookup: (title: string, artist: string) => Promise<LookupResult>;
+  lookup: (title: string, artist: string) => Promise<MetadataLookupResult>;
   enrichTracks: (
-    tracks: EnrichInputTrack[],
+    tracks: EnrichTrackInput[],
     options: { writeToFile: boolean; onlyMissing: boolean }
-  ) => Promise<EnrichResult[]>;
+  ) => Promise<EnrichTrackResult[]>;
   /**
    * Look-up-only single-track enrichment. Returns the would-be `updatedFields`
    * (and a cached cover URL when one was downloaded) WITHOUT writing tags or
@@ -57,18 +23,11 @@ export interface MetadataApi {
    * with code `metadata.enrich_busy` when a bulk run holds the abort slot.
    */
   previewEnrich: (
-    track: EnrichInputTrack,
+    track: EnrichTrackInput,
     options: { onlyMissing: boolean }
-  ) => Promise<EnrichResult>;
+  ) => Promise<EnrichTrackResult>;
   cancelEnrichment: () => Promise<void>;
-  onEnrichProgress: (
-    callback: (data: {
-      current: number;
-      total: number;
-      trackName: string;
-      status: 'searching' | 'downloading' | 'writing' | 'done' | 'error' | 'cancelled';
-    }) => void
-  ) => () => void;
+  onEnrichProgress: (callback: (data: EnrichProgress) => void) => () => void;
 }
 
 export const metadataApi: MetadataApi = {
@@ -76,10 +35,5 @@ export const metadataApi: MetadataApi = {
   enrichTracks: (tracks, options) => ipcRenderer.invoke(C.enrichTracks, tracks, options),
   previewEnrich: (track, options) => ipcRenderer.invoke(C.enrichPreview, track, options),
   cancelEnrichment: () => ipcRenderer.invoke(C.enrichCancel),
-  onEnrichProgress: createIpcListener<{
-    current: number;
-    total: number;
-    trackName: string;
-    status: 'searching' | 'downloading' | 'writing' | 'done' | 'error' | 'cancelled';
-  }>(C.enrichProgress),
+  onEnrichProgress: createIpcListener<EnrichProgress>(C.enrichProgress),
 };
