+++ apps/web/vite.config.ts
@@ -9,6 +9,7 @@ export default defineConfig({
   resolve: {
     alias: {
       '@': resolve(__dirname, './src'),
+      '@shiranami/contracts': resolve(__dirname, '../../packages/contracts/src/index.ts'),
       '@shiranami/shared': resolve(__dirname, '../../packages/shared/src/index.ts'),
     },
   },
