+++ packages/contracts/src/index.ts
@@ -2,4 +2,5 @@
 // shared API DTOs.
 export * from './domain/track';
 export * from './ipc/channels';
+export * from './ipc/metadata';
 export * from './share/dto';
